step 10 lang rework
"simplified" random chests command: removed true/false setting in favor of dis/enabling it via the two numeric parameters.
This way Users hopefully aren't as confused how to config this feature.

Added better time parsing for commands

Changed Reward system to match the Module one

removed label from commands it was never used

drop item rewards that doesn't fit into inventory

+++ Minigames/src/main/java/au/com/mineauz/minigames/FloorDegenerator.java
@@ -20,7 +20,7 @@ public class FloorDegenerator {
     private final Location xSidePos2;
     private final Location zSideNeg2;
     private final Location zSidePos2;
-    private int timeDelay = 30;
+    private long timeDelay = 30;
     private Minigame mgm = null;
     private int taskID = -1;
